requirements/prod.txt
viam-sdk==0.4.1
pygame==2.5.2
numpy==1.24.4
opencv-python==4.9.0.80
Pillow==9.5.0
